// src/event-tracking/Types/Types.ts
export enum Feature {
    VisitProductDetailPage = 'Visit Detail Product',
    UserAddAuctionPrice = 'Add Auction Price',
    ApprovedAuctionProduct = 'Approved Auction Product'
}
export enum AttributeType {
    BrowserName="BrowserName",
    LocalTimestamp="LocalTimestamp",
    SessionId="SessionId",
    UserId = 'UserId',
    CategoryName = 'Attribute3',
    CategoryId = 'Attribute4',
    ProductId = 'Attribute1',
    ProductName = 'Attribute2',
    AuctionPrice = 'Attribute5'
}